Migrate recent commits GraphQl query to triagebot queries
