+++ src/github/queries/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod issue_with_comments;
 pub(crate) mod least_recently_reviewed;
+pub(crate) mod recent_commits;
 pub(crate) mod user_comments_in_org;
 pub(crate) mod user_contributions;
 pub(crate) mod user_info;
