+++ src/github/repos.rs
@@ -5,11 +5,10 @@ use super::UserId;
 
 use anyhow::Context;
 use bytes::Bytes;
-use chrono::{DateTime, FixedOffset, Utc};
+use chrono::{DateTime, FixedOffset};
 use itertools::Itertools;
 use octocrab::models::Author;
 use reqwest::StatusCode;
-use std::collections::HashSet;
 use tracing as log;
 
 // User
@@ -1199,181 +1198,6 @@ impl Repository {
             })))
             .await
             .with_context(|| format!("{} failed to create commit for tree {tree}", self.full_name))
-    }
-}
-
-// Recent commits
-
-pub struct RecentCommit {
-    pub title: String,
-    pub pr_num: Option<i32>,
-    pub oid: String,
-    pub committed_date: DateTime<Utc>,
-}
-
-impl Repository {
-    /// Returns a list of recent commits on the given branch.
-    ///
-    /// Returns results in the OID range `oldest` (exclusive) to `newest`
-    /// (inclusive).
-    pub async fn recent_commits(
-        &self,
-        client: &GithubClient,
-        branch: &str,
-        oldest: &str,
-        newest: &str,
-    ) -> anyhow::Result<Vec<RecentCommit>> {
-        // This is used to deduplicate the results (so that a PR with multiple
-        // commits will only show up once).
-        let mut prs_seen = HashSet::new();
-        let mut recent_commits = Vec::new(); // This is the final result.
-        use cynic::QueryBuilder;
-        use github_graphql::docs_update_queries::{
-            GitObject, RecentCommits, RecentCommitsArguments,
-        };
-
-        let mut args = RecentCommitsArguments {
-            branch,
-            name: self.name(),
-            owner: self.owner(),
-            after: None,
-        };
-        let mut found_newest = false;
-        let mut found_oldest = false;
-        // This simulates --first-parent. We only care about top-level commits.
-        // Unfortunately the GitHub API doesn't provide anything like that.
-        let mut next_first_parent = None;
-        // Search for `oldest` within 3 pages (300 commits).
-        for _ in 0..3 {
-            let query = RecentCommits::build(args.clone());
-            let data = client
-                .json::<cynic::GraphQlResponse<RecentCommits>>(
-                    client.post(&client.graphql_url).json(&query),
-                )
-                .await
-                .with_context(|| {
-                    format!(
-                        "{} failed to get recent commits branch={branch}",
-                        self.full_name
-                    )
-                })?;
-
-            if let Some(errors) = data.errors {
-                anyhow::bail!("There were graphql errors. {errors:?}");
-            }
-            let target = data
-                .data
-                .context("No data returned.")?
-                .repository
-                .context("No repository.")?
-                .ref_
-                .context("No ref.")?
-                .target
-                .context("No target.")?;
-            let GitObject::Commit(commit) = target else {
-                anyhow::bail!("unexpected target type {target:?}")
-            };
-            let commits = commit
-                .history
-                .nodes
-                .into_iter()
-                // Don't include anything newer than `newest`
-                .skip_while(|node| {
-                    if found_newest || node.oid.0 == newest {
-                        found_newest = true;
-                        false
-                    } else {
-                        // This should only happen if there is a commit that arrives
-                        // between the time that `update_submodules` fetches the latest
-                        // ref, and this runs. This window should be a few seconds, so it
-                        // should be unlikely. This warning is here in case my assumptions
-                        // about how things work is not correct.
-                        tracing::warn!(
-                            "unexpected race with submodule history, newest oid={newest} skipping oid={}",
-                            node.oid.0
-                        );
-                        true
-                    }
-                })
-                // Skip nodes that aren't the first parent
-                .filter(|node| {
-                    let this_first_parent = node.parents.nodes
-                        .first()
-                        .map(|parent| parent.oid.0.clone());
-
-                    if let Some(first_parent) = &next_first_parent {
-                        if first_parent == &node.oid.0 {
-                            // Found the next first parent, include it and
-                            // set next_first_parent to look for this
-                            // commit's first parent.
-                            next_first_parent = this_first_parent;
-                            true
-                        } else {
-                            // Still looking for the next first parent.
-                            false
-                        }
-                    } else {
-                        // First commit.
-                        next_first_parent = this_first_parent;
-                        true
-                    }
-                })
-                // Stop once reached the `oldest` commit
-                .take_while(|node| {
-                    if node.oid.0 == oldest {
-                        found_oldest = true;
-                        false
-                    } else {
-                        true
-                    }
-                })
-                .filter_map(|node| {
-                    // Determine if this is associated with a PR or not.
-                    match node.associated_pull_requests
-                        // Get the first PR (we only care about one)
-                        .and_then(|mut pr| pr.nodes.pop()) {
-                        Some(pr) => {
-                            // Only include a PR once
-                            if prs_seen.insert(pr.number) {
-                                Some(RecentCommit {
-                                    pr_num: Some(pr.number),
-                                    title: pr.title,
-                                    oid: node.oid.0.clone(),
-                                    committed_date: node.committed_date,
-                                })
-                            } else {
-                                None
-                            }
-                        }
-                        None => {
-                            // This is an unassociated commit, possibly
-                            // created without a PR.
-                            Some(RecentCommit {
-                                pr_num: None,
-                                title: node.message_headline,
-                                oid: node.oid.0,
-                                committed_date: node.committed_date,
-                            })
-                        }
-                    }
-                });
-            recent_commits.extend(commits);
-            let page_info = commit.history.page_info;
-            if found_oldest || !page_info.has_next_page || page_info.end_cursor.is_none() {
-                break;
-            }
-            args.after = page_info.end_cursor;
-        }
-        if !found_oldest {
-            // This should probably do something more than log a warning, but
-            // I don't think it is too important at this time (the log message
-            // is only informational, and this should be unlikely to happen).
-            tracing::warn!(
-                "{} failed to find oldest commit sha={oldest} branch={branch}",
-                self.full_name
-            );
-        }
-        Ok(recent_commits)
     }
 }
